refactor(branch-resolver): extract domains into modular structure
Extract branch-resolver.js (835 LOC) into domain-driven modules:
- validation.js: Branch existence and content validation
- discovery.js: Git branch discovery and selection
- fallback.js: Post-merge fallback logic
- file-operations.js: Git file operations without checkout
- db-operations.js: Database operations for SD tracking

Key improvements:
- Windows cross-platform compatibility (replaced Unix shell syntax)
- All 26 tests passing
- Main file reduced from 835 to ~215 LOC (re-exports only)
